Add files via upload
Swing is a GUI widget toolkit for Java. It is part of Oracle's Java Foundation Classes (JFC) – an API for providing a graphical user interface (GUI) for Java programs. Example Swing widgets in Java. Swing was developed to provide a more sophisticated set of GUI components than the earlier Abstract Window Toolkit (AWT).